+++ app/src/components/mainWindow.ts
@@ -155,7 +155,7 @@ function createContextMenu(
 
 export function saveAppArgs(newAppArgs: OutputOptions): void {
   try {
-    fs.writeFileSync(APP_ARGS_FILE_PATH, JSON.stringify(newAppArgs));
+    fs.writeFileSync(APP_ARGS_FILE_PATH, JSON.stringify(newAppArgs, null, 2));
   } catch (err: unknown) {
     log.warn(
       `WARNING: Ignored nativefier.json rewrital (${(err as Error).message})`,